Add files via upload
Aim: To determine the factors of a number n and then return the pth element of the list, sorted ascending. If there is no pth element, return 0.
Note: To optimize the code, instead of iterating up to n, I iterate up to the square root of n, as the factors above the square root will be paired with the factors below the square root.